Implemented repeat request after sec

+++ data/src/main/kotlin/com/ringoid/data/remote/model/BaseResponse.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 open class BaseResponse(
     @Expose @SerializedName(COLUMN_ERROR_CODE) val errorCode: String = "",
     @Expose @SerializedName(COLUMN_ERROR_MESSAGE) val errorMessage: String = "",
-    @Expose @SerializedName(COLUMN_REPEAT_AFTER_SEC) val repeatAfterSec: Int = 0) {
+    @Expose @SerializedName(COLUMN_REPEAT_AFTER_SEC) val repeatAfterSec: Long = 0L) {
 
     companion object {
         const val COLUMN_ERROR_CODE = "errorCode"
